+++ convert_test.go
@@ -20,25 +20,23 @@ func TestConvert(t *testing.T) {
 	vbuf := nk.NewBuffer()
 	defer vbuf.Free()
 	configBuilder := nk.ConvertConfigBuilder{
-		CConvertConfig: nk.CConvertConfig{
-			VertexSize:      20, // sizeof(sdl.Vertex)
-			VertexAlignment: 4,  // alignof(sdl.Vertex)
-			Null: nk.DrawNullTexture{
-				Texture: nk.Handle(unsafe.Pointer(nil)),
-				UV:      nk.Vec2{X: 0, Y: 0},
-			},
-			CircleSegmentCount: 10,
-			CurveSegmentCount:  10,
-			ArcSegmentCount:    10,
-			GlobalAlpha:        1.0,
-			ShapeAA:            nk.AntiAliasingOff,
-			LineAA:             nk.AntiAliasingOff,
+		GlobalAlpha:        1.0,
+		ShapeAA:            nk.AntiAliasingOff,
+		LineAA:             nk.AntiAliasingOff,
+		CircleSegmentCount: 10,
+		ArcSegmentCount:    10,
+		CurveSegmentCount:  10,
+		Null: nk.DrawNullTexture{
+			Texture: nk.Handle(unsafe.Pointer(nil)),
+			UV:      nk.Vec2{X: 0, Y: 0},
 		},
 		VertexLayout: []nk.DrawVertexLayoutElement{
 			{Attribute: nk.VertexPosition, Format: nk.FormatFloat, Offset: 0},
 			{Attribute: nk.VertexTexcoord, Format: nk.FormatFloat, Offset: 8},
 			{Attribute: nk.VertexColor, Format: nk.FormatR8G8B8A8, Offset: 16},
 		},
+		VertexSize:      20, // sizeof(sdl.Vertex)
+		VertexAlignment: 4,  // alignof(sdl.Vertex)
 	}
 	config := configBuilder.Build()
 	defer config.Free()